Update metadata feature
This commit fixes the wrongly triggered warning popups if a mp game
is loaded and also adds the metadata to mp games.
The second changes prevents wrongly popups in normal MP games… (it is
hard to distinct between PB games and normal MP games at this stage of
loading.)

Note the trick to use 'GetLastPing(0)' as indicatior for MP games. Other
functions aren't return the wished information at this stage of loading.